Added CommonInterface for solvers; general cleanup
